fix: Transform branding properties to portletPreferences for sidebarLogin - meeds-io/MIPs#203 - MEED-9487
Before this fix, siebarLogin portlet use Branding information for background image, alt text, title and subtitle
With this it is not possible to update the background for a portlet without changing for all portlets.

This commit transforms branding properties into portletPreferences

+++ component/core/src/main/java/org/exoplatform/social/core/attachment/AttachmentServiceImpl.java
@@ -360,6 +360,39 @@ public class AttachmentServiceImpl implements AttachmentService {
                               String destinationObjectId,
                               String destinationParentObjectId,
                               long userIdentityId) {
+    moveAttachments(sourceObjectType,
+                    sourceObjectId,
+                    destinationObjectType,
+                    destinationObjectId,
+                    destinationParentObjectId,
+                    userIdentityId,
+                    true);
+  }
+
+  @Override
+  public void copyAttachments(String sourceObjectType,
+                              String sourceObjectId,
+                              String destinationObjectType,
+                              String destinationObjectId,
+                              String destinationParentObjectId,
+                              long userIdentityId) {
+    moveAttachments(sourceObjectType,
+                    sourceObjectId,
+                    destinationObjectType,
+                    destinationObjectId,
+                    destinationParentObjectId,
+                    userIdentityId,
+                    false);
+  }
+
+
+
+  private void moveAttachments(String sourceObjectType,
+                              String sourceObjectId,
+                              String destinationObjectType,
+                              String destinationObjectId,
+                              String destinationParentObjectId,
+                              long userIdentityId, boolean deleteAfterMove) {
     ObjectAttachmentList objectAttachmentList = getAttachments(sourceObjectType, sourceObjectId);
     List<ObjectAttachmentDetail> attachments = objectAttachmentList.getAttachments();
     if (CollectionUtils.isNotEmpty(attachments)) {
@@ -376,21 +409,23 @@ public class AttachmentServiceImpl implements AttachmentService {
                            destinationParentObjectId,
                            userIdentityId,
                            properties);
-          List<MetadataItem> metadataItemToDelete =
-                                                  metadataService.getMetadataItemsByMetadataNameAndTypeAndObject(attachment.getId(),
-                                                                                                                 AttachmentService.METADATA_TYPE.getName(),
-                                                                                                                 sourceObjectType,
-                                                                                                                 sourceObjectId,
-                                                                                                                 0,
-                                                                                                                 0);
-          if (CollectionUtils.isNotEmpty(metadataItemToDelete)) {
-            metadataItemToDelete.forEach(metadataItem -> {
-              try {
-                metadataService.deleteMetadataItem(metadataItem.getId(), true);
-              } catch (ObjectNotFoundException e) {
-                LOG.error("Error when deleting metadata item", e);
-              }
-            });
+          if (deleteAfterMove) {
+            List<MetadataItem> metadataItemToDelete =
+                metadataService.getMetadataItemsByMetadataNameAndTypeAndObject(attachment.getId(),
+                                                                               AttachmentService.METADATA_TYPE.getName(),
+                                                                               sourceObjectType,
+                                                                               sourceObjectId,
+                                                                               0,
+                                                                               0);
+            if (CollectionUtils.isNotEmpty(metadataItemToDelete)) {
+              metadataItemToDelete.forEach(metadataItem -> {
+                try {
+                  metadataService.deleteMetadataItem(metadataItem.getId(), true);
+                } catch (ObjectNotFoundException e) {
+                  LOG.error("Error when deleting metadata item", e);
+                }
+              });
+            }
           }
         } catch (Exception e) {
           LOG.error("Error when creating attachment", e);
